Replaced double with float
Double processing libraries were taking a nice chunk of memory and we
don't need such precision here.

+++ Core/Inc/pid.h
@@ -11,11 +11,11 @@
 #include "main.h"
 
 typedef struct pid_values {
-	double max;
-	double min;
-	double Kp;
-	double Ki;
-	double Kd;
+	float max;
+	float min;
+	float Kp;
+	float Ki;
+	float Kd;
 	int16_t maxI;
 	int16_t minI;
 } pid_values_t;
@@ -23,15 +23,15 @@ typedef struct pid_values {
 extern volatile pid_values_t currentPID;
 
 void setupPIDFromStruct();
-void setupPID(double max, double min, double Kp, double Kd, double Ki, int16_t _minI, int16_t _maxI );
-double calculatePID( double setpoint, double pv );
+void setupPID(float max, float min, float Kp, float Kd, float Ki, int16_t _minI, int16_t _maxI );
+float calculatePID( float setpoint, float pv );
 void resetPID();
-double getError();
-double getIntegral();
-double getPID_D();
-double getPID_P();
-double getPID_I();
-double getOutput();
-double getPID_SetPoint();
-double getPID_PresentValue();
+float getError();
+float getIntegral();
+float getPID_D();
+float getPID_P();
+float getPID_I();
+float getOutput();
+float getPID_SetPoint();
+float getPID_PresentValue();
 #endif /* PID_H_ */
